Use other prettyprint js lib
